Extract createMediaStreamTrackControllersFromMediaStream into independent utility
